+++ src/locales/en/location/index.ts
@@ -12,8 +12,8 @@ import continent from './continent';
 import country from './country';
 import county from './county';
 import direction from './direction';
-import full_address from './full_address';
 import language from './language';
+import postal_address from './postal_address';
 import postcode from './postcode';
 import secondary_address from './secondary_address';
 import state from './state';
@@ -33,8 +33,8 @@ const location: LocationDefinition = {
   country,
   county,
   direction,
-  full_address,
   language,
+  postal_address,
   postcode,
   secondary_address,
   state,
