+++ src/locales/ro_MD/location/index.ts
@@ -8,7 +8,7 @@ import city_name from './city_name';
 import city_pattern from './city_pattern';
 import city_prefix from './city_prefix';
 import county from './county';
-import full_address from './full_address';
+import postal_address from './postal_address';
 import postcode from './postcode';
 import secondary_address from './secondary_address';
 import state from './state';
@@ -24,7 +24,7 @@ const location: LocationDefinition = {
   city_pattern,
   city_prefix,
   county,
-  full_address,
+  postal_address,
   postcode,
   secondary_address,
   state,
