+++ src/locales/zh_CN/location/index.ts
@@ -11,8 +11,8 @@ import city_suffix from './city_suffix';
 import continent from './continent';
 import country from './country';
 import direction from './direction';
-import full_address from './full_address';
 import language from './language';
+import postal_address from './postal_address';
 import postcode from './postcode';
 import secondary_address from './secondary_address';
 import state from './state';
@@ -31,8 +31,8 @@ const location: LocationDefinition = {
   continent,
   country,
   direction,
-  full_address,
   language,
+  postal_address,
   postcode,
   secondary_address,
   state,
