+++ src/modules/location/index.ts
@@ -368,16 +368,16 @@ export class LocationModule extends SimpleLocationModule {
    * Generates a random localized full postal address, which may include a street address, secondary address, city, state, and zip code. To ensure you get locale-specific address formats, use a localized Faker instance.
    *
    * @example
-   * faker.location.fullAddress() // 'Apt. 980\n0917 O'Conner Estates\nWest Shannonview\nMichigan\n82180'
-   * fakerEN_US.location.fullAddress() // '0917 O'Conner Estates, Apt. 980\nWest Shannonview, MI 82180'
-   * fakerEN_GB.location.fullAddress() // '79 Bogan Corner\nCastle Zemlakborough\nDumfries and Galloway\nZH17 2SD'
-   * fakerZH_CN.location.fullAddress() // '广东省贵原市门路19号'
+   * faker.location.postalAddress() // 'Apt. 980\n0917 O'Conner Estates\nWest Shannonview\nMichigan\n82180'
+   * fakerEN_US.location.postalAddress() // '0917 O'Conner Estates, Apt. 980\nWest Shannonview, MI 82180'
+   * fakerEN_GB.location.postalAddress() // '79 Bogan Corner\nCastle Zemlakborough\nDumfries and Galloway\nZH17 2SD'
+   * fakerZH_CN.location.postalAddress() // '广东省贵原市门路19号'
    *
    * @since 10.4.0
    */
-  fullAddress(): string {
+  postalAddress(): string {
     return this.faker.helpers.fake(
-      this.faker.definitions.location.full_address
+      this.faker.definitions.location.postal_address
     );
   }
